requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
discord.py==1.7.3
idna==3.3
mimicbot-chat==1.0.0
multidict==6.0.2
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
typer==0.6.1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
